refactor(monitor): 纯依赖注入，职责边界清晰
- 删除 BiliDanmakuMonitor.create() 静态工厂方法
- BiliDanmakuMonitor 成为纯依赖注入类，不管理连接生命周期
- MonitorTaskWorker 管理 BiliApiClient 生命周期，注入给 Monitor
- DanmakuVerifier 添加 create() 工厂方法和 verify_by_cid/verify_all_pending 类方法
- HistoryController 直接使用 DanmakuVerifier
