+++ src/danmaku_sender/service/bili_monitor.py
@@ -1,15 +1,11 @@
 """弹幕监视器 - 负责持续盯住某个 Target 并出统计指标"""
 
 import logging
-from contextlib import contextmanager
-from typing import Callable
 
 from .danmaku_verifier import DanmakuVerifier
 
-from danmaku_sender.repo.bili_api_client import BiliApiClient
 from danmaku_sender.repo.history_manager import HistoryManager
-from danmaku_sender.types.models.common import VideoTarget, MonitorStats, VerifyResult
-from danmaku_sender.config import ApiAuthConfig
+from danmaku_sender.types.models.common import VideoTarget, MonitorStats
 
 
 logger = logging.getLogger(__name__)
@@ -21,20 +17,14 @@ class BiliDanmakuMonitor:
 
     负责针对特定 VideoTarget 的持续监视周期管理。
     内部使用 DanmakuVerifier 执行核销操作。
+
+    纯依赖注入设计：不管理连接生命周期，由调用方负责创建和注入依赖。
     """
 
     def __init__(self, verifier: DanmakuVerifier, target: VideoTarget, history_manager: HistoryManager):
         self.verifier = verifier
         self.target = target
         self.history_manager = history_manager
-
-    @staticmethod
-    @contextmanager
-    def create(target: VideoTarget, auth_config: ApiAuthConfig, history_manager: HistoryManager):
-        """Context manager 工厂：自动管理 client 生命周期"""
-        with BiliApiClient.from_config(auth_config) as client:
-            verifier = DanmakuVerifier(api_client=client, history_manager=history_manager)
-            yield BiliDanmakuMonitor(verifier=verifier, target=target, history_manager=history_manager)
 
     def monitor(self, stats_baseline: float = 0.0) -> MonitorStats:
         """
@@ -69,43 +59,3 @@ class BiliDanmakuMonitor:
             'pending': pending,
             'lost': lost
         }
-
-    @classmethod
-    def verify_by_cid(cls, cid: int, auth_config: ApiAuthConfig, history_manager: HistoryManager) -> VerifyResult:
-        """
-        轻量级单次验证：拉取指定 CID 的在线弹幕，核销存活并标记丢失。
-
-        Returns:
-            VerifyResult: {'verified': int, 'lost': int, 'total_checked': int}
-        """
-        with BiliApiClient.from_config(auth_config) as client:
-            verifier = DanmakuVerifier(api_client=client, history_manager=history_manager)
-            return verifier.verify_cid(cid)
-
-    @classmethod
-    def verify_all_pending(
-        cls,
-        auth_config: ApiAuthConfig,
-        history_manager: HistoryManager,
-        on_progress: Callable[[int, int, VerifyResult], None] | None = None
-    ) -> VerifyResult:
-        """
-        批量验证所有含有待验证弹幕的 CID。
-
-        Args:
-            on_progress: 进度回调 (已处理数, 总数, 本次增量结果)
-
-        Returns:
-            VerifyResult: {'verified': int, 'lost': int, 'total_checked': int}
-        """
-        pending_cids = history_manager.get_pending_cids()
-
-        if not pending_cids:
-            logger.info("没有待验证的弹幕记录。")
-            return {'verified': 0, 'lost': 0, 'total_checked': 0}
-
-        cids = [entry['cid'] for entry in pending_cids]
-
-        with BiliApiClient.from_config(auth_config) as client:
-            verifier = DanmakuVerifier(api_client=client, history_manager=history_manager)
-            return verifier.verify_batch(cids, on_progress=on_progress)
